Controller implementation done

+++ src/main/kotlin/pt/unl/fct/iadi/bookstore/controller/dto/FindBookResponse.kt
@@ -1,6 +1,7 @@
 package pt.unl.fct.iadi.bookstore.controller.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
+import java.math.BigDecimal
 
 data class FindBookResponse(
     @field:Schema(description = "Identifier of the book")
@@ -13,7 +14,7 @@ data class FindBookResponse(
     val author: String,
 
     @field:Schema(description = "Price of the book")
-    val price: Double,
+    val price: BigDecimal,
 
     @field:Schema(description = "Remote address (URL) to an image of the book cover")
     val image: String
